#4 Refactor the code base
