+++ src/component/Navbar.js
@@ -1,14 +1,13 @@
 import React, {useContext} from "react";
-import logo from "./Instagrill.png";
+import logo from "../Instagrill.png";
 import HomeRoundedIcon from "@material-ui/icons/HomeRounded";
 import ExploreOutlinedIcon from "@material-ui/icons/ExploreOutlined";
 import FavoriteBorderRoundedIcon from "@material-ui/icons/FavoriteBorderRounded";
 import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import SearchIcon from '@material-ui/icons/Search';
 import Popover from '@material-ui/core/Popover';
 import { makeStyles } from '@material-ui/core/styles';
 import { Link } from "react-router-dom";
-import {firebaseAuth} from './provider/AuthProvider';
+import {firebaseAuth} from '../provider/AuthProvider';
 
 const useStyles = makeStyles((theme) => ({
   popover: {
