+++ src/component/Post.js
@@ -1,10 +1,9 @@
 import React, { useState, useEffect } from "react";
-import "./index.css";
+import "../index.css";
 import Avatar from "@material-ui/core/Avatar";
 import FavoriteBorder from "@material-ui/icons/FavoriteBorder";
 import ChatBubbleOutlineOutlinedIcon from "@material-ui/icons/ChatBubbleOutlineOutlined";
-// import data from './data/database.json';
-import { database } from './firebase';
+import { database } from '../firebase/firebase';
 
 const Post = () => {
   const [likeCounter, setLikeCounter] = useState(0);
